+++ web-client/src/components/forms/Button/Button.rendering.test.tsx
@@ -10,4 +10,20 @@ describe("Button.component: rendering", () => {
     expect(buttonElement).not.toBeNull()
     expect(buttonElement.innerHTML).toContain(label)
   })
+
+  describe("when 'svgOnly' is set as intent", () => {
+    it("renders the button with specified icon but no visible label", () => {
+      const label = "Click me"
+      render(
+        <Button intent={"svgOnly"} iconName={"dots"}>
+          {label}
+        </Button>
+      )
+
+      const buttonElement = screen.getByRole("button")
+      expect(buttonElement).not.toBeNull()
+      expect(buttonElement.innerHTML).toContain(label)
+      expect(buttonElement.children[1].className).contains("sr-only")
+    })
+  })
 })