+++ web-client/src/components/layout/menu/Sidebar/Sidebar.rendering.test.tsx
@@ -1,0 +1,14 @@
+import { nanoid } from "nanoid"
+import { render, screen } from "@testing-library/react"
+import { Sidebar } from "@/components/layout/menu"
+import { boards } from "@/utils/mocks/boards.mocks"
+
+describe("Sidebar.component: rendering", () => {
+  it("renders the component", () => {
+    const testId = nanoid()
+    render(<Sidebar testId={testId} boards={boards} />)
+
+    const sidebarElement = screen.getByTestId(testId)
+    expect(sidebarElement).not.toBeNull()
+  })
+})