+++ web-client/src/components/layout/menu/ThemeSwitcher/ThemeSwitcher.behaviour.test.tsx
@@ -46,4 +46,58 @@ describe("ThemeSwitcher.component: behavior", () => {
       })
     })
   })
+
+  describe("when the user choose keyboard for the state change", () => {
+    describe("and the 'Enter' key is pressed", () => {
+      it("toggles the state", () => {
+        const testId = nanoid()
+        const onThemeUpdate = vitest.fn()
+
+        render(
+          <ThemeSwitcher
+            testId={testId}
+            default="light"
+            onThemeUpdate={onThemeUpdate}
+          />
+        )
+
+        const themeSwitcherElement = screen.getByTestId(testId)
+
+        fireEvent.keyDown(themeSwitcherElement, { key: "Enter" })
+        fireEvent.keyDown(themeSwitcherElement, { key: "Enter" })
+        fireEvent.keyDown(themeSwitcherElement, { key: "Enter" })
+
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(1, "dark")
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(2, "light")
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(3, "dark")
+      })
+    })
+
+    describe("and the 'Left and Right Arrow' keys are pressed", () => {
+      it("toggles the state", () => {
+        const testId = nanoid()
+        const onThemeUpdate = vitest.fn()
+
+        render(
+          <ThemeSwitcher
+            testId={testId}
+            default="dark"
+            onThemeUpdate={onThemeUpdate}
+          />
+        )
+
+        const themeSwitcherElement = screen.getByTestId(testId)
+
+        fireEvent.keyDown(themeSwitcherElement, { key: "ArrowRight" })
+        fireEvent.keyDown(themeSwitcherElement, { key: "ArrowRight" })
+        fireEvent.keyDown(themeSwitcherElement, { key: "ArrowRight" })
+        fireEvent.keyDown(themeSwitcherElement, { key: "ArrowLeft" })
+
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(1, "light")
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(2, "dark")
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(3, "light")
+        expect(onThemeUpdate).toHaveBeenNthCalledWith(4, "dark")
+      })
+    })
+  })
 })