+++ web-client/src/components/layout/navigation/Logo/Logo.component.tsx
@@ -5,8 +5,8 @@ import LogoWide from "@/assets/logo-wide.svg"
  */
 export const Logo = () => {
   return (
-    <div className="flex flex-col justify-center w-75 h-24 bg-custom-white pl-8 border-r-2">
-      <a href="/">
+    <div className="flex flex-col justify-center w-75 bg-custom-white pl-8 border-r-2">
+      <a href="/web-client/public">
         <LogoWide />
       </a>
     </div>