+++ web-client/src/components/layout/navigation/index.ts
@@ -1,0 +1,2 @@
+export { Logo } from "./Logo/Logo.component"
+export { Navbar } from "./Navbar/Navbar.component"