+++ web-client/src/utils/mocks/menus.mocks.ts
@@ -1,0 +1,23 @@
+import { nanoid } from "nanoid"
+
+type ItemType = "standard" | "destructive"
+
+export type BoardContextMenu = {
+  id: string
+  title: string
+  type: ItemType
+  onClickHandler?: (type: ItemType) => void
+}
+
+export const boardContextMenuItems: BoardContextMenu[] = [
+  {
+    id: nanoid(),
+    title: "Edit Board",
+    type: "standard",
+  },
+  {
+    id: nanoid(),
+    title: "Delete Board",
+    type: "destructive",
+  },
+]